changed the INT to BIGINT in create and insert folders

+++ sql_scripts/create/event.sql
@@ -1,6 +1,6 @@
 DROP TABLE IF EXISTS event;
 CREATE TABLE event(
-    id_event INT PRIMARY KEY NOT NULL AUTO_INCREMENT,
+    id_event BIGINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
     title_event TINYTEXT,
     seat_event TINYTEXT,
     date_start_event DATE,
